Tried a lot of things with rendering
. Attempted to add shadows
. Fixed a bug with model library loading models that didn't exist
. Made it possible to draw 2D quad meshes w/textures
. Made it possible to add custom ImGui display modules for components
. Removed a lot of old code
. Found glaring issue with blend modes
TODO: Go to object on selection, drag entities into folders and sub folders, disable movement while paused or not focused in game window, selectable entities need to have their identifiers be Their ID so we can click on them, use WIC texture loader for images for potentially better alpha results

+++ GlowEngine/Source/Engine/Graphics/Window/Window.cpp
@@ -14,8 +14,8 @@
 // construct the window
 Graphics::Window::Window()
   : 
-  windowWidth(GetSystemMetrics(SM_CXSCREEN)/1.25f),
-  windowHeight(GetSystemMetrics(SM_CYSCREEN)/1.25f)
+  windowWidth((float)GetSystemMetrics(SM_CXSCREEN)/1.25f),
+  windowHeight((float)GetSystemMetrics(SM_CYSCREEN)/1.25f)
 {
   windowClassName = L"Otherglow Window";
   windowName = L"Otherglow";
